+++ user-service/src/main/java/com/waither/userservice/entity/UserMedian.java
@@ -1,9 +1,12 @@
 package com.waither.userservice.entity;
 
+import com.waither.userservice.converter.SurveyConverter;
 import com.waither.userservice.entity.enums.Season;
 import com.waither.userservice.global.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.util.List;
 
 import static com.waither.userservice.global.util.CalculateUtil.calculateMedian;
 
@@ -29,12 +32,35 @@ public class UserMedian extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private Season season;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
 
     public void setUser(User user) {
         this.user = user;
+    }
+
+    // UserMedian 기본값으로 설정
+    public static UserMedian createUserMedian(UserData userData) {
+        return UserMedian.builder()
+                .medianOf1And2(calculateMedian(userData.getLevel1(), userData.getLevel2()))
+                .medianOf2And3(calculateMedian(userData.getLevel2(), userData.getLevel3()))
+                .medianOf3And4(calculateMedian(userData.getLevel3(), userData.getLevel4()))
+                .medianOf4And5(calculateMedian(userData.getLevel4(), userData.getLevel5()))
+                .season(userData.getSeason())
+                .build();
+    }
+
+    // UserMedianList 생성
+    public static List<UserMedian> createUserMedianList(List<UserData> userDataList, User user) {
+        return userDataList.stream()
+                .map(userData -> {
+                    UserMedian userMedian = createUserMedian(userData);
+                    // 연관관계 설정
+                    userMedian.setUser(user);
+                    return userMedian;
+                })
+                .toList();
     }
 
     public void updateMedianValue(UserData userData) {
